refactor: renamed MainActivityClickHandler to HomeFragmentClickHandler and moved to the appropriate package
